Третий блок завершен. Осталось: Токен и Уведомления

+++ app/src/main/java/models/Ticket.kt
@@ -2,7 +2,7 @@ package models
 
 class Ticket (
     val ticket_id: Long,
-    val flight_number: Int,
+    val flight_number: Long,
     val time_start: String,
     val seat_number: Int,
     val code_number: String
